requirements.txt
pywebview>=5.0.0
pandas>=2.0.0
openpyxl>=3.1.0
email-validator>=2.1.0
